+++ pom.xml
@@ -5,7 +5,7 @@
 <modelVersion>4.0.0</modelVersion>
 <groupId>com.gmail.st3venau.plugins</groupId>
 <artifactId>ArmorStandTools</artifactId>
-<version>4.4.0</version>
+<version>4.4.1</version>
 <name>ArmorStandTools</name>
 
 <repositories>
